Moves wagmi config into separate file
